DTSPB-4088 fix coding Style issue
